requirements-dev.txt
-r requirements.txt
pytest-playwright>=0.5.0
playwright>=1.44.0
